Change options system
Now youtubedlg stores the non-string part of the options
into settings.json. This way we only need a TwoWayOrderedDict
on the optionsframe.py module to store the options(video_format,
 audio_quality, etc..) instead of using a dictionary on the parsers.py
module and a list with all the available values on the optionsframe.py
module. This change also helps us to keep all the GUI strings on the
GUI modules making it easier to translate the GUI.

NOTE
=====
Youtubedlg will reset all the stored options on the first usage of the
new system due to compatibility issues with the old settings.json.

CHANGES
=======

optionsframe.py
  Store the values into a TwoWayOrderedDict and map their values-keys
  on the load-save_options methods.

optionsmanager.py
  Enable the value check system on the _settings_are_valid() method.

parsers.py
  Remove all the module attributes and some of the options_build methods.