Locale is set automatically
Language is set automatically during the first run. If the locale can not be detected, the locale is set to en. If the locale is detected but the .mo file is not included in youtube-dl-gui, the locale is set to en. An user can still change language manually in settings. Tested in Mint 18.2 and Win 7.